CSCService: include device name only in scan response
Switched out the device name to be only included when responding to a scan request.
This saves on advertising space as we are limited to 31 octets, so now only the
profiles (and powerlevel) are included in the single advertising data, which then
triggers a response data when another device queries it.  This reolves long
Bluetooth devices names as this would otherwise trigger a
ADVERTISE_FAILED_DATA_TOO_LARGE error code.