Fix stopping BLE bridge
When stopping the BLE bridge, the CSCService was not being stopped and thus the device
would continue to poll ANT and also would have some odd close behaviours.

Updated the code to remember the intent that is created for the service, and also to
call the unbind for the service when stopping, as well as taking ownership of the tasks
for the service that the app generates by updating the manifest.  Now when stopping
the BLE bridge, the app actually stops the service that it creates on start up.